+++ sample-dApps/game-contract/src/logic/tests.rs
@@ -1,6 +1,6 @@
 use super::*;
 use naumachia::address::Address;
-use naumachia::ledger_client::in_memory_ledger::TestBackendsBuilder;
+use naumachia::ledger_client::test_ledger_client::TestBackendsBuilder;
 use naumachia::smart_contract::{SmartContract, SmartContractTrait};
 
 #[tokio::test]
