+++ src/ledger_client/test_ledger_client.rs
@@ -1,4 +1,3 @@
-use std::fs::File;
 use std::{
     fmt::Debug,
     hash::Hash,
@@ -7,22 +6,25 @@ use std::{
 };
 use uuid::Uuid;
 
-use crate::ledger_client::local_persisted_ledger::LocalPersistedStorage;
-use crate::ledger_client::{build_outputs, new_wallet_output};
-use crate::transaction::TxId;
+use crate::ledger_client::test_ledger_client::in_memory_storage::InMemoryStorage;
 use crate::{
     backend::Backend,
     ledger_client::LedgerClientError::FailedToIssueTx,
-    ledger_client::{LedgerClient, LedgerClientError, LedgerClientResult},
+    ledger_client::{build_outputs, new_wallet_output},
+    ledger_client::{LedgerClient, LedgerClientResult},
     output::Output,
+    transaction::TxId,
     values::Values,
     Address, PolicyId, UnbuiltTransaction,
 };
 use async_trait::async_trait;
-use serde::de::DeserializeOwned;
-use serde::Serialize;
+use local_persisted_storage::LocalPersistedStorage;
+use serde::{de::DeserializeOwned, Serialize};
 use tempfile::TempDir;
 use thiserror::Error;
+
+pub mod in_memory_storage;
+pub mod local_persisted_storage;
 
 pub struct TestBackendsBuilder<Datum, Redeemer> {
     signer: Address,
@@ -105,8 +107,6 @@ where
     }
 }
 
-type MutableData<Datum> = Arc<Mutex<Vec<(Address, Output<Datum>)>>>;
-
 #[derive(Debug, Error)]
 enum InMemoryLCError {
     #[error("Mutex lock error: {0:?}")]
@@ -128,83 +128,6 @@ pub trait TestLedgerStorage<Datum> {
     async fn all_outputs(&self, address: &Address) -> LedgerClientResult<Vec<Output<Datum>>>;
     async fn remove_output(&self, output: &Output<Datum>) -> LedgerClientResult<()>;
     async fn add_output(&self, output: &Output<Datum>) -> LedgerClientResult<()>;
-}
-
-#[derive(Debug)]
-pub struct InMemoryStorage<Datum> {
-    signer: Address,
-    outputs: MutableData<Datum>,
-}
-
-#[async_trait::async_trait]
-impl<Datum: Clone + Send + Sync + PartialEq> TestLedgerStorage<Datum> for InMemoryStorage<Datum> {
-    async fn signer(&self) -> LedgerClientResult<Address> {
-        Ok(self.signer.clone())
-    }
-
-    async fn outputs_by_count(
-        &self,
-        address: &Address,
-        count: usize,
-    ) -> LedgerClientResult<Vec<Output<Datum>>> {
-        let outputs = self
-            .outputs
-            .lock()
-            .map_err(|e| InMemoryLCError::Mutex(format! {"{:?}", e}))
-            .map_err(|e| {
-                LedgerClientError::FailedToRetrieveOutputsAt(address.clone(), Box::new(e))
-            })?
-            .iter()
-            .cloned()
-            .filter_map(|(a, o)| if &a == address { Some(o) } else { None })
-            .take(count)
-            .collect();
-        Ok(outputs)
-    }
-
-    async fn all_outputs(&self, address: &Address) -> LedgerClientResult<Vec<Output<Datum>>> {
-        let outputs = self
-            .outputs
-            .lock()
-            .map_err(|e| InMemoryLCError::Mutex(format! {"{:?}", e}))
-            .map_err(|e| {
-                LedgerClientError::FailedToRetrieveOutputsAt(address.clone(), Box::new(e))
-            })?
-            .iter()
-            .cloned()
-            .filter_map(|(a, o)| if &a == address { Some(o) } else { None })
-            .collect();
-        Ok(outputs)
-    }
-
-    async fn remove_output(&self, output: &Output<Datum>) -> LedgerClientResult<()> {
-        let mut ledger_utxos = self
-            .outputs
-            .lock()
-            .map_err(|e| InMemoryLCError::Mutex(format! {"{:?}", e}))
-            .map_err(|e| FailedToIssueTx(Box::new(e)))?;
-        let index = ledger_utxos
-            .iter()
-            .position(|(_, x)| x == output)
-            .ok_or_else(|| {
-                LedgerClientError::FailedToRetrieveOutputWithId(
-                    output.id().clone(),
-                    Box::new(InMemoryLCError::DuplicateInput),
-                )
-            })?;
-        ledger_utxos.remove(index);
-        Ok(())
-    }
-
-    async fn add_output(&self, output: &Output<Datum>) -> LedgerClientResult<()> {
-        let mut ledger_utxos = self
-            .outputs
-            .lock()
-            .map_err(|e| InMemoryLCError::Mutex(format! {"{:?}", e}))
-            .map_err(|e| FailedToIssueTx(Box::new(e)))?;
-        ledger_utxos.push((output.owner().clone(), output.clone()));
-        Ok(())
-    }
 }
 
 #[derive(Debug)]
@@ -321,91 +244,5 @@ where
             self.storage.add_output(&output).await?;
         }
         Ok(TxId::new("Not a real id"))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #![allow(non_snake_case)]
-    use super::*;
-    use crate::ledger_client::local_persisted_ledger::starting_output;
-    use crate::output::UnbuiltOutput;
-    use tempfile::TempDir;
-
-    #[tokio::test]
-    async fn outputs_at_address() {
-        let tmp_dir = TempDir::new().unwrap();
-        let path = tmp_dir.path().join("data");
-        let signer = Address::new("alice");
-        let starting_amount = 10_000_000;
-        // let record = TestLedgerClient::<(), ()>::init(&path, signer.clone(), starting_amount).unwrap();
-        let output = starting_output::<()>(&signer, starting_amount);
-        let outputs = vec![(signer.clone(), output)];
-        let record: TestLedgerClient<(), (), _> =
-            TestLedgerClient::new_in_memory(signer.clone(), outputs);
-        let mut outputs = record.all_outputs_at_address(&signer).await.unwrap();
-        assert_eq!(outputs.len(), 1);
-        let first_output = outputs.pop().unwrap();
-        let expected = starting_amount;
-        let actual = first_output.values().get(&PolicyId::ADA).unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[tokio::test]
-    async fn balance_at_address() {
-        let tmp_dir = TempDir::new().unwrap();
-        let path = tmp_dir.path().join("data");
-        let signer = Address::new("alice");
-        let starting_amount = 10_000_000;
-        let output = starting_output::<()>(&signer, starting_amount);
-        let outputs = vec![(signer.clone(), output)];
-        let record: TestLedgerClient<(), (), _> =
-            TestLedgerClient::new_in_memory(signer.clone(), outputs);
-        let expected = starting_amount;
-        let actual = record
-            .balance_at_address(&signer, &PolicyId::ADA)
-            .await
-            .unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[tokio::test]
-    async fn issue() {
-        let tmp_dir = TempDir::new().unwrap();
-        let path = tmp_dir.path().join("data");
-        let signer = Address::new("alice");
-        let starting_amount = 10_000_000;
-        let output = starting_output::<()>(&signer, starting_amount);
-        let outputs = vec![(signer.clone(), output)];
-        let record: TestLedgerClient<(), (), _> =
-            TestLedgerClient::new_in_memory(signer.clone(), outputs);
-        let first_output = record
-            .all_outputs_at_address(&signer)
-            .await
-            .unwrap()
-            .pop()
-            .unwrap();
-        let owner = Address::new("bob");
-        let new_output = UnbuiltOutput::new_wallet(owner.clone(), first_output.values().clone());
-        let tx: UnbuiltTransaction<(), ()> = UnbuiltTransaction {
-            script_inputs: vec![],
-            unbuilt_outputs: vec![new_output],
-            minting: Default::default(),
-            policies: Default::default(),
-        };
-        record.issue(tx).await.unwrap();
-        let expected_bob = starting_amount;
-        let actual_bob = record
-            .balance_at_address(&owner, &PolicyId::ADA)
-            .await
-            .unwrap();
-        assert_eq!(expected_bob, actual_bob);
-
-        let expected_alice = 0;
-        let actual_alice = record
-            .balance_at_address(&signer, &PolicyId::ADA)
-            .await
-            .unwrap();
-        assert_eq!(expected_alice, actual_alice)
     }
 }
